feat(dashboard): show multi-tenant throughput series and legend checkboxes
- Replace single-tenant dropdown with direct multi-tenant series comparison on the throughput trend chart.
- Support ChartLegend toggling per-tenant and aggregated curves with metric toggle (RPM/TPM).
- Split DashboardPage into cohesive modules (DashboardKpiCard, DashboardMetrics, ThroughputTrendChart) to satisfy file size gates.
